Moving a basic message builder to the core library.

+++ src/RapideFix/Parsers/MessageParser.cs
@@ -133,7 +133,7 @@ namespace RapideFix.Parsers
       ReadOnlySpan<byte> messagePart = message;
       int indexSOH = 0;
       int indexEquals = 0;
-      int checksumValue = 0;
+      byte checksumValue = 0;
       while(messagePart.Length > checksumLength)
       {
         if(TraverseMessageBody(messagePart, out indexEquals, out indexSOH, ref checksumValue))
@@ -156,11 +156,11 @@ namespace RapideFix.Parsers
       return targetObject;
     }
 
-    private bool TraverseMessageBody(ReadOnlySpan<byte> messagePart, out int indexEquals, out int indexSOH, ref int checksumValue)
+    private bool TraverseMessageBody(ReadOnlySpan<byte> messagePart, out int indexEquals, out int indexSOH, ref byte checksumValue)
     {
       indexEquals = -1;
       indexSOH = -1;
-      int checksum = 0;
+      byte checksum = 0;
       int i;
       for(i = 0; i < messagePart.Length; i++)
       {
